make MainNode fully abstract
